+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **BREAKING:** Require PHP 8 and Laravel 8.
+- **BREAKING:** Store credentials in database instead of using JWT tokens.
+
 
 ## [3.0.0] - 2020-12-09