+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [4.0.0] - 2021-06-09
+
 ### Changed
 - **BREAKING:** Require PHP 8 and Laravel 8.
 - **BREAKING:** Store credentials in database instead of using JWT tokens.